Text fixture loading.
